requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
coverage==7.4.0
frozenlist==1.4.1
furl==2.1.3
idna==3.6
iniconfig==2.0.0
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
magic-filter==1.0.12
multidict==6.0.4
multidimensional-urlencode==0.0.4
nose==1.3.7
orderedmultidict==1.0.1
packaging==23.2
pluggy==1.3.0
pook==1.4.0
pydantic==2.5.3
pydantic_core==2.14.6
pytest==7.4.4
pytest-asyncio==0.23.3
python-dotenv==1.0.0
referencing==0.32.1
requests==2.31.0
retailcrm==5.1.2
rpds-py==0.17.1
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
xmltodict==0.13.0
yarl==1.9.4
